Fix cuda out of memory issue
PyTorch creates a computational graph whenever we pass
data through the network and stores the computations on the GPU
memory, in case we want to calculate the gradient during backpropagation.
When we only want to perform a forward propagation, specify torch.no_grad().

In short: during eval phases, specify "with torch.no_grad()"

Here during eval phase, for every possible maneuver we were running
through a forward pass of a Transformer: do not store data for backprop
in that case...